requirements.txt
attrs==22.1.0
Automat==20.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
constantly==15.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jedi==0.18.1
Naked==0.1.31
packaging==21.3
parso==0.8.3
pudb==2022.1.2
pycryptodome==3.15.0
Pygments==2.13.0
pyparsing==3.0.9
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
six==1.16.0
Twisted==22.4.0
typing-extensions==4.3.0
urllib3==1.26.12
urwid==2.1.2
urwid-readline==0.13
zope.interface==5.4.0
